Move picking pixels into separate file
